Extract all the referral logic in a concern
